+++ src/routes/(authenticated)/+layout.server.ts
@@ -4,7 +4,7 @@ import type { LayoutServerLoad } from "./$types";
 export const load: LayoutServerLoad = async ({ locals: { supabase, isLogged, getSession }, }) => {
     const loginCheck = isLogged();
 
-    console.log(await getSession())
+    console.log("AWW")
     if (loginCheck !== "has auth") redirect(302, "/");
