+++ src/routes/(authenticated)/+layout.svelte
@@ -1,33 +1,7 @@
 <script lang="ts">
 	import Nav from '$lib/auth-components/nav.svelte';
-	import { getAuthState, setAuthState, setUserState } from '$lib';
+	import { getAuthState } from '$lib';
 	import ProjectJoinedContent from '$lib/auth-components/projects/join-project/project-joined-content.svelte';
-	import type { LayoutServerData } from '../$types';
-
-	export let data: LayoutServerData;
-
-	const defaultAuthState = {
-		activeItem: '/dashboard',
-		projects: {
-			joinedProject: false,
-			showEditTools: false
-		},
-		guilds: {
-			joinedGuild: false,
-			guildObj: {
-				imageUrl: '',
-				hostName: '',
-				maxUsers: 0,
-				guildName: '',
-				guildDescription: '',
-				isPrivate: false,
-				joinedCount: 0
-			}
-		}
-	};
-
-	$: data.session?.user ? setUserState(data.session.user) : setUserState(null);
-	setAuthState(defaultAuthState);
 
 	const authState = getAuthState();
 </script>
