+++ src/routes/+layout.server.ts
@@ -1,6 +1,7 @@
 import type { LayoutServerLoad } from './$types'
 
 export const load: LayoutServerLoad = async ({ locals: { getSession } }) => {
+    console.log("INVALIDATED")
     return {
         session: await getSession(),
     }
