+++ src/routes/+page.server.ts
@@ -42,12 +42,6 @@ export const actions: Actions = {
             const { data: { user }, error: registerError } = await supabase.auth.signUp({
                 email: result.email,
                 password: result.password,
-                options: {
-                    data: {
-                        firstname: result.firstName,
-                        lastname: result.lastName,
-                    }
-                }
             });
 
             if (registerError) return fail(401, { msg: registerError.message });
@@ -55,14 +49,13 @@ export const actions: Actions = {
                 const { error: insertError } = await supabase.from("user_list_tb").insert([{
                     user_id: user.id,
                     user_email: user.email,
-                    user_firstname: user.user_metadata.firstname,
-                    user_lastname: user.user_metadata.lastname
+                    user_fullname: `${result.lastName}, ${result.firstName}`,
                 }]);
 
                 if (insertError) {
                     const { error: deleteUserError } = await supabaseAdmin.auth.admin.deleteUser(user.id);
                     if (deleteUserError) return fail(401, { msg: deleteUserError.message });
-                    else return fail(401, { msg: "There is an error try again." });
+                    else return fail(401, { msg: insertError.message });
                 } else return fail(200, { msg: "Registered Successfully." });
             }
 
@@ -89,22 +82,19 @@ export const actions: Actions = {
         if (checkLogin === "has auth" && session) {
             try {
                 const result = updateInformationSchema.parse(formData);
-                const { data: { user }, error: updateUserError } = await supabase.auth.updateUser({
-                    data: {
-                        firstname: result.firstName,
-                        lastname: result.lastName,
-                        bio: result.bio,
-                        address: result.address,
-                        barangay: result.barangay,
-                        city: result.city,
-                        religion: result.religion,
-                        contact: result.contactNumber,
-                        profileLink: session.user.user_metadata.profileLink
-                    }
-                });
+
+                const { error: updateUserError } = await supabase.from("user_list_tb").update([{
+                    user_fullname: `${result.lastName}, ${result.firstName}`,
+                    user_bio: result.bio,
+                    user_address: result.address,
+                    user_barangay: result.barangay,
+                    user_city: result.city,
+                    user_religion: result.religion,
+                    user_contact: result.contactNumber
+                }])
 
                 if (updateUserError) return fail(401, { msg: updateUserError.message });
-                else if (user) return fail(200, { msg: 'Information Updated Successfully.', user });
+                else return fail(200, { msg: 'Information Updated Successfully.' });
 
 
             } catch (error) {
@@ -134,17 +124,15 @@ export const actions: Actions = {
             else if (uploadPicture) {
                 const { data: { publicUrl } } = supabase.storage.from("collab-bucket").getPublicUrl(uploadPicture.path)
 
-                const { data: { user }, error: updateUserError } = await supabase.auth.updateUser({
-                    data: {
-                        profileLink: `${publicUrl}?${Math.random()}`
-                    }
-                });
+                const { error: updateUserError } = await supabase.from("user_list_tb").update([{
+                    user_photo_link: `${publicUrl}?${Math.random()}`
+                }]).eq("user_id", session.user.id);
 
                 if (updateUserError) {
                     //this is alternative atm transaction comming soon
                     await supabase.storage.from("collab-bucket").remove([session.user.id])
                     return fail(401, { msg: updateUserError.message });
-                } else if (user) return fail(200, { msg: "Upload successfully", user })
+                } else return fail(200, { msg: "Upload successfully" });
             }
 
         } else redirect(302, "/");
